Remove problematic comparison between str and int
In Python 2 it is "legal" to compare object of different types, though
rules are far from obvious.  In Python 3.x the behavior has been changed
so that attempting to order an integer and a string will raise an error.

I removed the problematic if statement as set_block_info is always
called in blockpre.py anyway (Thanks Michelle for pointing that out!)